Weather-Agent AI/pyproject.toml
[tool.poetry]
name = "weather-agent"
version = "0.1.0"
description = "Simple Weather Agent"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
crewai = "0.1.24"
unstructured = "0.10.25"
pyowm = "3.3.0"
python-dotenv = "1.0.0"
langchain-openai = "0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
